+++ lib/src/webview_method_channel.dart
@@ -22,8 +22,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
 
   final MethodChannel _channel;
 
-  static const MethodChannel _cookieManagerChannel =
-      MethodChannel('plugins.flutter.io/cookie_manager');
+  static const MethodChannel _cookieManagerChannel = MethodChannel('plugins.flutter.io/cookie_manager');
 
   Future<dynamic> _onMethodCall(MethodCall call) async {
     switch (call.method) {
@@ -43,13 +42,31 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
       case 'onPageStarted':
         _platformCallbacksHandler.onPageStarted(call.arguments['url']);
         return null;
+      case 'onWebResourceError':
+        _platformCallbacksHandler.onWebResourceError(
+          WebResourceError(
+            errorCode: call.arguments['errorCode'],
+            description: call.arguments['description'],
+            domain: call.arguments['domain'],
+            errorType: call.arguments['errorType'] == null
+                ? null
+                : WebResourceErrorType.values.firstWhere(
+                    (WebResourceErrorType type) {
+                      return type.toString() == '$WebResourceErrorType.${call.arguments['errorType']}';
+                    },
+                  ),
+          ),
+        );
+        return null;
       case "onJsBridgeCall":
         return _platformCallbacksHandler.onJsBridgeCall(call);
       case "dismissLoadingMask":
         return _platformCallbacksHandler.dismissLoadingMask();
     }
+
     throw MissingPluginException(
-        '${call.method} was invoked but has no handler');
+      '${call.method} was invoked but has no handler',
+    );
   }
 
   @override
@@ -96,20 +113,17 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
 
   @override
   Future<String> evaluateJavascript(String javascriptString) {
-    return _channel.invokeMethod<String>(
-        'evaluateJavascript', javascriptString);
+    return _channel.invokeMethod<String>('evaluateJavascript', javascriptString);
   }
 
   @override
   Future<void> addJavascriptChannels(Set<String> javascriptChannelNames) {
-    return _channel.invokeMethod<void>(
-        'addJavascriptChannels', javascriptChannelNames.toList());
+    return _channel.invokeMethod<void>('addJavascriptChannels', javascriptChannelNames.toList());
   }
 
   @override
   Future<void> removeJavascriptChannels(Set<String> javascriptChannelNames) {
-    return _channel.invokeMethod<void>(
-        'removeJavascriptChannels', javascriptChannelNames.toList());
+    return _channel.invokeMethod<void>('removeJavascriptChannels', javascriptChannelNames.toList());
   }
 
   @override
@@ -129,9 +143,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
 
   /// Method channel mplementation for [WebViewPlatform.clearCookies].
   static Future<bool> clearCookies() {
-    return _cookieManagerChannel
-        .invokeMethod<bool>('clearCookies')
-        .then<bool>((dynamic result) => result);
+    return _cookieManagerChannel.invokeMethod<bool>('clearCookies').then<bool>((dynamic result) => result);
   }
 
   static Map<String, dynamic> _webSettingsToMap(WebSettings settings) {
@@ -153,8 +165,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
     _addIfNonNull('jsMode', settings.javascriptMode?.index);
     _addIfNonNull('hasNavigationDelegate', settings.hasNavigationDelegate);
     _addIfNonNull('debuggingEnabled', settings.debuggingEnabled);
-    _addIfNonNull(
-        'gestureNavigationEnabled', settings.gestureNavigationEnabled);
+    _addIfNonNull('gestureNavigationEnabled', settings.gestureNavigationEnabled);
     _addSettingIfPresent('userAgent', settings.userAgent);
     return map;
   }
@@ -163,8 +174,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
   ///
   /// This is used for the `creationParams` argument of the platform views created by
   /// [AndroidWebViewBuilder] and [CupertinoWebViewBuilder].
-  static Map<String, dynamic> creationParamsToMap(
-      CreationParams creationParams) {
+  static Map<String, dynamic> creationParamsToMap(CreationParams creationParams) {
     return <String, dynamic>{
       'initialUrl': creationParams.initialUrl,
       'settings': _webSettingsToMap(creationParams.webSettings),